requirements.txt
fastapi
uvicorn[standard]
pandas
sqlalchemy
psycopg2-binary
chromadb
llama-index
scikit-learn
python-dotenv
pyyaml
pillow
python-multipart
requests
streamlit
langsmith
PyPDF2
pytesseract
convert
pillow
pytesseract
openpyxl
huggingface_hub>=0.14.1
